% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Function to download specific datasets used throughout the book}
\usage{
get_data(dataset)
}
\arguments{
\item{dataset}{The name of the dataset. Should be one of "benchmarks",
"ds-bowl-2019", "k-entry", "sesame",  "state-test", "titanic", or
"violence"}
}
\description{
Function to download specific datasets used throughout the book
}
