% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdsr-package.R
\docType{package}
\name{sds-package}
\alias{sds}
\alias{sds-package}
\title{sds: Package to accompany the Social Data Science book}
\description{
Functions to fetch data used in examples in the book. Also includes associated learnr apps.
}
\author{
\strong{Maintainer}: Daniel Anderson \email{daniela@uoregon.edu} (\href{https://orcid.org/0000-0003-4699-4680}{ORCID})

Authors:
\itemize{
  \item Brendan Cullen \email{bcullen@uoregon.edu} (\href{https://orcid.org/0000-0003-4825-132X}{ORCID})
  \item Ouafaa Hmaddi \email{ohmaddi@uoregon.edu} (\href{https://orcid.org/0000-0003-1443-4830}{ORCID})
}

}
\keyword{internal}
