% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{tic_seconds}
\alias{tic_seconds}
\title{tictoc seconds}
\usage{
tic_seconds(log_n)
}
\arguments{
\item{log_n}{An integer specifying the number from \code{tictoc::tic.log}}
}
\description{
Formats the output from tictoc timings so they can be more easily included in text
}
